style: ...

+++ src/main/kotlin/xyz/cssxsh/bilibili/data/Live.kt
@@ -7,7 +7,7 @@ interface Live {
     val roomId: Long
     val title: String
     val cover: String
-    val online: Int
+    val online: Long
     val liveStatus: Boolean
     val link: String
 }
@@ -35,37 +35,25 @@ data class BiliRoomInfo(
     @SerialName("lock_till")
     @Serializable(NumberToBooleanSerializer::class)
     val lockTill: Boolean,
-//    @SerialName("need_p2p")
-//    @Serializable(NumberToBooleanSerializer::class)
-//    private val needP2P: Boolean,
     @SerialName("pwd_verified")
     val passwordVerified: Boolean,
     @SerialName("room_id")
     val roomId: Long,
-//    @SerialName("room_shield")
-//    @Serializable(NumberToBooleanSerializer::class)
-//    private val roomShield: Boolean,
     @SerialName("short_id")
     val shortId: Int,
-//    @SerialName("special_type")
-//    private val specialType: Int,
     @SerialName("uid")
     val uid: Long
 )
 
 @Serializable
 data class BiliRoomSimple(
-//    @SerialName("broadcast_type")
-//    private val broadcastType: Int,
     @SerialName("cover")
     override val cover: String,
-//    @SerialName("link")
-//    val link: String? = null,
     @SerialName("liveStatus")
     @Serializable(NumberToBooleanSerializer::class)
     override val liveStatus: Boolean,
     @SerialName("online")
-    override val online: Int,
+    override val online: Long,
     @SerialName("online_hidden")
     @Serializable(NumberToBooleanSerializer::class)
     val onlineHidden: Boolean = false,
@@ -117,51 +105,21 @@ data class BiliLiveOff(
 
 @Serializable
 data class LiveRecommend(
-//    @SerialName("area_v2_id")
-//    val areaId: Int,
-//    @SerialName("area_v2_name")
-//    val areaName: String,
-//    @SerialName("area_v2_parent_id")
-//    val areaParentId: Int,
-//    @SerialName("area_v2_parent_name")
-//    val areaParentName: String,
-//    @SerialName("broadcast_type")
-//    private val broadcastType: Int,
-//    @SerialName("click_callback")
-//    private val clickCallback: String,
     @SerialName("cover")
     override val cover: String,
     @SerialName("face")
     val face: String,
     @SerialName("flag")
     val flag: Int,
-//    @SerialName("global_order")
-//    private val globalOrder: Int,
-//    @SerialName("group_id")
-//    private val groupId: Int,
-//    @SerialName("head_box")
-//    private val headBox: JsonElement? = null,
-//    @SerialName("head_box_type")
-//    private val headBoxType: Int,
     @SerialName("is_auto_play")
     @Serializable(NumberToBooleanSerializer::class)
     val isAutoPlay: Boolean,
-//    @SerialName("keyframe")
-//    private val keyframe: String,
     @SerialName("link")
     override val link: String,
     @SerialName("online")
-    override val online: Int,
-//    @SerialName("pendant_Info")
-//    private val pendantInfo: JsonElement? = null,
+    override val online: Long,
     @SerialName("roomid")
     override val roomId: Long,
-//    @SerialName("session_id")
-//    private val sessionId: String,
-//    @SerialName("show_callback")
-//    private val showCallback: String,
-//    @SerialName("special_id")
-//    private val specialId: Long,
     @SerialName("title")
     override val title: String,
     @SerialName("uid")
@@ -171,8 +129,6 @@ data class LiveRecommend(
     @SerialName("liveStatus")
     @Serializable(NumberToBooleanSerializer::class)
     override val liveStatus: Boolean = true,
-//    @SerialName("verify")
-//    private val verify: JsonElement? = null
 ) : Live
 
 @Serializable
@@ -181,8 +137,6 @@ data class LiveRecord(
     val bvid: String,
     @SerialName("cover")
     val cover: String,
-//    @SerialName("global_order")
-//    private val globalOrder: Int,
     @SerialName("is_sticky")
     @Serializable(NumberToBooleanSerializer::class)
     val isSticky: Boolean,
